Create likes.js module
