+++ src/index.js
@@ -1,12 +1,13 @@
 import './index.css';
 import './popup.css';
 import './loader.css';
-import { renderShowsToDOM, sendLikeToAPI } from './modules/renderToDOM.js';
+import { renderShowsToDOM } from './modules/renderToDOM.js';
+import { sendLikeToAPI } from './modules/likes.js';
 import { createPopup } from './modules/popup.js';
 
 const showsContainer = document.querySelector('.maze__grid');
 
-showsContainer.addEventListener('click', (e) => {
+showsContainer.addEventListener('click', async (e) => {
   const commentBtn = e.target.closest('.comment__btn');
   const likeBtn = e.target.closest('.like__btn');
 
@@ -17,7 +18,10 @@ showsContainer.addEventListener('click', (e) => {
 
   if (likeBtn) {
     const { id } = likeBtn.closest('.maze__card');
-    sendLikeToAPI(id);
+    const likeState = await sendLikeToAPI(id);
+
+    if (likeState) {
+    }
   }
 });
